+++ src/Deriver/Runner.fs
@@ -3,9 +3,9 @@ namespace Lmc.KafkaApplication.Deriver
 module internal DeriverRunner =
     open Lmc.KafkaApplication
 
-    let runDeriver: RunPattern<DeriverApplication<'InputEvent, 'OutputEvent>, 'InputEvent, 'OutputEvent> =
+    let runDeriver: RunPattern<DeriverApplication<'InputEvent, 'OutputEvent, 'Dependencies>, 'InputEvent, 'OutputEvent, 'Dependencies> =
         fun run (DeriverApplication application) ->
-            let beforeRun _: BeforeRun<'InputEvent, 'OutputEvent> =
+            let beforeRun _: BeforeRun<'InputEvent, 'OutputEvent, 'Dependencies> =
                 ignore
 
             application