+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add generic dependencies to runtime parts
 
 ## 21.2.0 - 2022-08-26
 - Update dependencies